+++ src/api/core/_credentialsFactory.ts
@@ -1,8 +1,8 @@
 import { AxiosResponse, default as axios } from 'axios';
 import { access, constants, readFileSync, writeFileSync } from 'fs';
 import { dirname } from 'path';
-import { AcountNumberString, Credentials, IAccount, IAccounts, ITime, QuestradeAPIOptions } from '../../typescript';
 import { sync } from '../../utils/mkdirp';
+import { AcountNumberString, Credentials, IAccount, IAccounts, ITime, QuestradeAPIOptions } from './typescript';
 import { _axiosApiGet } from './_axiosApi';
 
 const _getServerTime = (credentials: Credentials) => async () =>
