+++ src/api/core/_getDataFromApi.ts
@@ -1,6 +1,6 @@
-import { Credentials, IAccounts, IActivities, IBalances, ICandle, ICandles, IExecutions, IMarkets, IOrders, IPositions, IQuotes, ISymbol, ISymbols, Time } from '../../typescript';
-import { IOptionChains } from '../../typescript/IOptionsQuotes';
 import { endpointFormatDateTool } from '../../utils/timeutil';
+import { Credentials, IAccounts, IActivities, IBalances, ICandle, ICandles, IExecutions, IMarkets, IOrders, IPositions, IQuotes, ISymbol, ISymbols, Time } from './typescript';
+import { IOptionChains } from './typescript/IOptionsQuotes';
 import { _axiosAccountApi as _axiosAccountGetApi, _axiosApiGet as _axiosGetApi } from './_axiosApi';
 // import { endpointFormatDateTool as _endpointFormatDate } from '../../../utils/endpointFormatDate';
