+++ src/api/core/typescript/IQuotes.ts
