+++ src/api/core/typescript/IStockSymbol.ts
