+++ src/api/core/typescript/ISymbols.ts
